New folders and some files are refactored (setup.py is deleted). A new way to save plots.
