+++ source/core/handling_msg.py
@@ -2,13 +2,15 @@
 In this module we process events related to bot (such as messages, requests)
 """
 
+from pathlib import Path
+
 from telegram import Update
 from telegram.ext import CallbackContext
 
-from graph import Graph
-from parser import Parser, ParseError
+from source.math.graph import Graph
+from source.math.parser import Parser, ParseError
 
-FILE_NAME = 'graph.png'
+RESOURCES_PATH = Path('resources')
 
 
 def echo(text: str):
@@ -19,18 +21,19 @@ def echo(text: str):
 def send_graph(update: Update, context: CallbackContext):
     """User requested to draw a plot"""
     user = update.message.from_user
+    file_path = RESOURCES_PATH / f"{user['id']}.png"
     expr = " ".join(context.args)
     parser = Parser()
 
     try:
         tokens = parser.parse(expr)
-        graph = Graph(FILE_NAME)
+        graph = Graph(file_path)
         graph.draw(tokens)
     except ParseError as err:
         update.message.reply_text(str(err))
         return
 
-    with open(FILE_NAME, 'rb') as graph_file:
+    with open(file_path, 'rb') as graph_file:
         output_message = "Here a graph of requested functions"
         if warning := parser.pop_last_warning():
             output_message += f"\n{warning}"
